Migrate Mutation.updateProjectName
